+++ crates/kitsune_p2p/dht_arc/src/lib.rs
@@ -1,5 +1,20 @@
+mod defaults;
+pub use defaults::*;
+
 mod dht_arc;
 pub use dht_arc::*;
+
+mod dht_arc_bucket;
+pub use dht_arc_bucket::*;
+
+mod dht_arc_redundancy;
+pub use dht_arc_redundancy::*;
+
+mod dht_arc_set;
+pub use dht_arc_set::*;
+
+mod peer_view;
+pub use peer_view::*;
 
 #[cfg(any(test, feature = "test_utils"))]
 pub mod loc8;
