+++ crates/kitsune_p2p/dht_arc/src/loc8.rs
@@ -1,6 +1,6 @@
 use std::collections::BTreeSet;
 
-use crate::dht_arc::{loc_downscale, loc_upscale, ArcInterval, DhtLocation};
+use crate::{loc_downscale, loc_upscale, ArcInterval, DhtLocation};
 
 /// A representation of DhtLocation in the u8 space. Useful for writing tests
 /// that test the full range of possible locations while still working with small numbers.
