+++ crates/kitsune_p2p/dht_arc/src/peer_view.rs
@@ -1,0 +1,14 @@
+mod peer_view_alpha;
+pub use peer_view_alpha::*;
+
+pub mod gaps;
+
+#[derive(Debug, Clone)]
+pub enum PeerStrat {
+    Alpha(PeerStratAlpha),
+}
+
+#[derive(Debug, Clone, derive_more::From)]
+pub enum PeerView {
+    Alpha(PeerViewAlpha),
+}
