+++ crates/kitsune_p2p/dht_arc/src/test/mod.rs
@@ -3,4 +3,5 @@ pub use ascii::ascii;
 
 mod arc_stability;
 mod intersection;
+mod test_ascii;
 mod union;
