+++ crates/kitsune_p2p/dht_arc/src/test/test_ascii.rs
@@ -1,4 +1,5 @@
-use super::*;
+use crate::ArcInterval;
+
 use pretty_assertions::assert_eq;
 
 #[test]
